# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (public checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      security:
        - adminToken: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, processing, shipped, delivered, cancelled]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status or notes (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete an order (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Customers ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OrderItem:
      type: object
      required: [productId, name, image, price, quantity]
      properties:
        productId:
          type: integer
        name:
          type: string
        image:
          type: string
        price:
          type: number
        quantity:
          type: integer
    Order:
      type: object
      required:
        [
          id, orderNumber, customerName, customerEmail, customerPhone,
          address, status, items, subtotal, shippingCost, tax,
          discountCode, discountAmount, total, notes, createdAt
        ]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        tax:
          type: number
        discountCode:
          type: ["string", "null"]
        discountAmount:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [customerName, customerEmail, customerPhone, address, items]
      properties:
        customerName:
          type: string
          minLength: 1
        customerEmail:
          type: string
          minLength: 1
        customerPhone:
          type: string
        address:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        discountCode:
          type: string
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        notes:
          type: string

    # ── Customer schemas ─────────────────────────────────────────────────────
...