# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  customerEmail: string;
  customerPhone: string;
  address: string;
  /** @minItems 1 */
  items: OrderItem[];
  discountCode?: string;
  notes?: string;
}