# lib/api-zod/src/generated/types/orderUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderUpdateStatus } from './orderUpdateStatus';

export interface OrderUpdate {
  status?: OrderUpdateStatus;
  notes?: string;
}